// frontend-admin/src/api/generated/model/generateReceiptRequest.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API â€” RKSV-compliant POS / Admin / shared Auth surfaces
 * OpenAPI spec version: v1
 */
import type { CompanyInfo } from './companyInfo';
import type { CustomerInfo } from './customerInfo';
import type { ReceiptItem } from './receiptItem';

export interface GenerateReceiptRequest {
  companyInfo?: CompanyInfo;
  customerInfo?: CustomerInfo;
  /** @nullable */
  items?: ReceiptItem[] | null;
  /**
   * @minLength 1
   * @maxLength 10
   */
  language: string;
  /** @nullable */
  paymentMethod?: string | null;
  taxAmount?: number;
  /**
   * @minLength 1
   * @maxLength 50
   */
  templateType: string;
  totalAmount?: number;
}